Add files via upload
Coalesce  : Gives the 1st non null value in the column.
Full join : Full Join creates a new table by joining two tables as a whole. The joined table contains all records from both the tables and fill in NULLs for missing matches on either side. In short, full join is a type of outer join that combines the results of both left and right joins.